Controllo per cancellazione bulk

+++ modules/ddt/bulk.php
@@ -145,14 +145,20 @@ switch (post('op')) {
     break;
 
     case 'delete-bulk':
+	
+		if ($debug){
+			
+			foreach ($id_records as $id) {
+				$dbo->query('DELETE  FROM dt_ddt  WHERE id = '.prepare($id).Modules::getAdditionalsQuery($id_module));
+				$dbo->query('DELETE FROM dt_righe_ddt WHERE idddt='.prepare($id).Modules::getAdditionalsQuery($id_module));
+				$dbo->query('DELETE FROM mg_movimenti WHERE idddt='.prepare($id).Modules::getAdditionalsQuery($id_module));
+			}
 
-        foreach ($id_records as $id) {
-            $dbo->query('DELETE  FROM dt_ddt  WHERE id = '.prepare($id).Modules::getAdditionalsQuery($id_module));
-            $dbo->query('DELETE FROM dt_righe_ddt WHERE idddt='.prepare($id).Modules::getAdditionalsQuery($id_module));
-            $dbo->query('DELETE FROM mg_movimenti WHERE idddt='.prepare($id).Modules::getAdditionalsQuery($id_module));
-        }
-
-        $_SESSION['infos'][] = tr('Ddt eliminati!');
+			$_SESSION['infos'][] = tr('Ddt eliminati!');
+			
+		}else{
+				$_SESSION['warnings'][] = tr('Procedura in fase di sviluppo. Nessuna modifica apportata.');
+		}
 
     break;
 }
